backend-fastapi/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0